# api-docs.json
{"openapi":"3.0.0","info":{"title":"Single Service API Docs","version":"1.0.0"},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"CreateBarangRequest":{"type":"object","required":["nama","harga","stok","perusahaan_id","kode"],"properties":{"nama":{"type":"string"},"harga":{"type":"number"},"stok":{"type":"number"},"perusahaan_id":{"type":"string"},"kode":{"type":"string"}}},"SuccessBarangResponse":{"type":"object","properties":{"status":{"type":"string","default":"success"},"message":{"type":"string"},"data":{"type":"object","properties":{"id":{"type":"string"},"nama":{"type":"string"},"harga":{"type":"number"},"stok":{"type":"number"},"perusahaan_id":{"type":"string"},"kode":{"type":"string"}}}}},"FailedToSearchBarangResponse":{"type":"object","properties":{"status":{"type":"string","default":"error"},"message":{"type":"string"},"data":{"type":null,"nullable":true}}},"CreatePerusahaanRequest":{"type":"object","required":["nama","harga","stok","perusahaan_id","kode"],"properties":{"nama":{"type":"string"},"alamat":{"type":"string"},"no_telp":{"type":"string"},"kode":{"type":"string"}}},"SuccessPerusahaanResponse":{"type":"object","properties":{"status":{"type":"string","default":"success"},"message":{"type":"string"},"data":{"type":"object","properties":{"nama":{"type":"string"},"alamat":{"type":"string"},"no_telp":{"type":"string"},"kode":{"type":"string"}}}}},"FailedToSearchPerusahaanResponse":{"type":"object","properties":{"status":{"type":"string","default":"error"},"message":{"type":"string"},"data":{"type":null,"nullable":true}}},"LoginRequest":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string","default":"admin"},"password":{"type":"string","default":"admin"}}},"LoginResponse":{"type":"object","properties":{"status":{"type":"string","default":"success"},"message":{"type":"string","default":"Session created"},"data":{"type":"object","properties":{"user":{"type":"object","properties":{"username":{"type":"string"},"name":{"type":"string"}}},"token":{"type":"string"}}}}},"InvalidCredentialsResponse":{"type":"object","properties":{"status":{"type":"string","default":"error"},"message":{"type":"string","default":"Invalid username or password"},"data":{"type":null,"nullable":true}}},"InvalidTokenResponse":{"type":"object","properties":{"status":{"type":"string","default":"error"},"message":{"type":"string","default":"Invalid token, no token"},"data":{"type":null,"nullable":true}}}}},"security":[{"BearerAuth":[]}],"paths":{"/self":{"get":{"tags":["Auth"],"description":"Get current info of auth","security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Success"},"401":{"description":"Has to have auth bearer token (via /login) to access this route"}}}},"/login":{"post":{"tags":["Auth"],"description":"Logging in user","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Has to have auth bearer token (via /login) to access this route","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidCredentialsResponse"}}}}}}},"/barang":{"post":{"tags":["Barang"],"description":"Adding barang to the database","security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBarangRequest"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessBarangResponse"}}}},"401":{"description":"Token Error / No Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidTokenResponse"}}}}}},"get":{"tags":["Barang"],"description":"Get all barang","parameters":[{"in":"query","name":"q","required":false,"type":"string"},{"in":"query","name":"perusahaan","required":false,"type":"string"}],"responses":{"200":{"description":"Sucess","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessBarangResponse"}}}},"500":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchResponse"}}}}}}},"/barang/{id}":{"get":{"tags":["Barang"],"description":"Get all barang","parameters":[{"name":"id","in":"path","description":"barang id","required":true}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessBarangResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchBarangResponse"}}}}}},"put":{"tags":["Barang"],"description":"Update barang to the database","parameters":[{"name":"id","in":"path","description":"barang id","required":true}],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBarangRequest"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessBarangResponse"}}}},"401":{"description":"Token Error / No Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidTokenResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchBarangResponse"}}}}}},"delete":{"tags":["Barang"],"description":"Delete barang from database","parameters":[{"name":"id","in":"path","description":"barang id","required":true}],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessBarangResponse"}}}},"401":{"description":"Token Error / No Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidTokenResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchBarangResponse"}}}}}}},"/perusahaan":{"post":{"tags":["Perusahaan"],"description":"Adding perusahaan to the database","security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePerusahaanRequest"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessPerusahaanResponse"}}}},"401":{"description":"Token Error / No Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidTokenResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchPerusahaanResponse"}}}}}},"get":{"tags":["Perusahaan"],"description":"Get all perusahaan","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessPerusahaanResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchPerusahaanResponse"}}}}}}},"/perusahaan/{id}":{"get":{"tags":["Perusahaan"],"description":"Get perusahaan detail by id","parameters":[{"name":"id","in":"path","description":"perusahaan id","required":true}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessPerusahaanResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchPerusahaanResponse"}}}}}},"put":{"tags":["Perusahaan"],"description":"Update perusahaan to the database","parameters":[{"name":"id","in":"path","description":"perusahaan id","required":true}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePerusahaanRequest"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessPerusahaanResponse"}}}},"401":{"description":"Token Error / No Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidTokenResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchPerusahaanResponse"}}}}}},"delete":{"tags":["Perusahaan"],"description":"Delete perusahaan from database","parameters":[{"name":"id","in":"path","description":"perusahaan id","required":true}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessPerusahaanResponse"}}}},"401":{"description":"Token Error / No Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvalidTokenResponse"}}}},"404":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailedToSearchPerusahaanResponse"}}}}}}}},"tags":[]}